+++ sat/cli/bootsys/util.py
@@ -25,55 +25,8 @@ OTHER DEALINGS IN THE SOFTWARE.
 from collections import defaultdict
 import logging
 import re
-import shlex
-import subprocess
-import time
 
 LOGGER = logging.getLogger(__name__)
-
-
-class RunningService:
-    def __init__(self, service, sleep_after_start=0):
-        """
-        Create a new context manager that starts a service in a context and
-        stops it when leaving the context.
-
-        Note that if a service was already running, it will be stopped after
-        the context is exited. It does not leave the service in the state it
-        was in before entering the context.
-
-        Args:
-            service (str): name of the service to start/stop
-            sleep_after_start (int): number of seconds to sleep after starting
-                the service when entering the context.
-        """
-        self._service = service
-        self._sleep_after_start = sleep_after_start
-
-    def _systemctl_start_stop(self, state):
-        """Start or stop the given service.
-
-        Args:
-            state (bool): If True, start the service. If False, stop it.
-
-        Returns:
-            None.
-        """
-        cmd = 'systemctl {} {}'.format('start' if state else 'stop',
-                                       self._service)
-        LOGGER.info('Running `%s`', cmd)
-        subprocess.check_call(shlex.split(cmd))
-
-    def __enter__(self):
-        self._systemctl_start_stop(True)
-
-        if self._sleep_after_start:
-            LOGGER.info("Sleeping for %s seconds after starting service %s.",
-                        self._sleep_after_start, self._service)
-            time.sleep(self._sleep_after_start)
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        self._systemctl_start_stop(False)
 
 
 # Maps from management NCN subroles to prefixes for those hostnames